Split the code into modules
