+++ bg_module.py
@@ -1,0 +1,23 @@
+import os
+import pygame
+
+# Background Image
+bg = pygame.image.load(os.path.join('Utils/Pics/Background','bg.png')).convert()
+bg_x = 0
+bg_width = bg.get_width()  
+background_speed = 2	# Background shifts by 2 pixels in each game loop
+
+def draw_bg(win):
+	global bg_x, bg_width
+	win.blit(bg, (bg_x, 0))
+	win.blit(bg, (bg_width,0))
+
+	# Background movement
+	bg_x -= background_speed
+	bg_width -= background_speed
+	
+	# For repetition of background movement
+	if bg_x < bg.get_width() * -1:
+		bg_x = bg.get_width()
+	if bg_width < bg.get_width() * -1:
+		bg_width = bg.get_width()
